move files to standardized ui.py/utils.py/components.py
instead of a very long single file with all of the logic
