+++ src/python/impactx/dashboard/Input/lattice/defaults_handler/__init__.py
@@ -1,0 +1,6 @@
+
+from .ui import LatticeDefaultsHandler
+
+__all__ = [
+    "LatticeDefaultsHandler",
+]
